refactor: reorganize MCP server structure
- Move server.rs from src/cli/cmd/mcp/ to src/mcp/
- Extract MCP tools into separate files in src/mcp/tools/:
  - search_scraps tool → tools/search_scraps.rs
  - list_tags tool → tools/list_tags.rs
- Update import paths throughout codebase
- Use modern Rust module structure without mod.rs files
- Maintain all functionality while improving code organization
