+++ src/cli/cmd/mcp.rs
@@ -1,2 +1,1 @@
 pub mod serve;
-pub mod server;
