+++ src/mcp/tools.rs
@@ -1,0 +1,2 @@
+pub mod list_tags;
+pub mod search_scraps;
